refactor:  View의 Input과 Output 분리
