+++ src/main/java/lotto/controller/Controller.java
@@ -7,14 +7,17 @@ import lotto.domain.LottoMachine;
 import lotto.domain.Prize;
 import lotto.domain.Prizes;
 import lotto.domain.WinningLotto;
-import lotto.view.View;
+import lotto.view.InputView;
+import lotto.view.OutputView;
 
 public class Controller {
-    private final View view;
+    private final InputView inputView;
+    private final OutputView outputView;
     private final LottoMachine lottoMachine;
 
-    public Controller(View view, LottoMachine lottoMachine) {
-        this.view = view;
+    public Controller(InputView inputView, OutputView outputView, LottoMachine lottoMachine) {
+        this.inputView = inputView;
+        this.outputView = outputView;
         this.lottoMachine = lottoMachine;
     }
 
@@ -26,29 +29,29 @@ public class Controller {
         Prizes prizes = new Prizes(lotto, winningLotto);
 
         printReward(prizes);
-        view.printRewardRate(prizes.getRewardRate(LottoMachine.LOTTO_PRICE * lottoCount));
+        outputView.printRewardRate(prizes.getRewardRate(LottoMachine.LOTTO_PRICE * lottoCount));
     }
 
     private List<Lotto> buyLotto() {
-        int expense = view.promptExpense();
+        int expense = inputView.promptExpense();
         List<Lotto> lotto = lottoMachine.issue(expense);
-        view.printLotto(lotto);
+        outputView.printLotto(lotto);
         return lotto;
     }
 
     private WinningLotto publishWinningLotto() {
-        List<Integer> numbers = view.promptWinningNumbers();
-        Integer bonus = view.promptBonusNumber();
+        List<Integer> numbers = inputView.promptWinningNumbers();
+        Integer bonus = inputView.promptBonusNumber();
         return new WinningLotto(numbers, bonus);
     }
 
     private void printReward(Prizes prizes) {
-        view.printPrizeHeader();
+        outputView.printPrizeHeader();
         Prize.reversedValuesForReward().forEach(prize -> printPrize(prize, prizes));
     }
 
     private void printPrize(Prize prize, Prizes prizes) {
         int prizeCount = prizes.count(prize);
-        view.printPrizeInfo(prize, prizeCount);
+        outputView.printPrizeInfo(prize, prizeCount);
     }
 }
