+++ src/main/java/lotto/view/OutputView.java
@@ -1,25 +1,11 @@
 package lotto.view;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.Scanner;
-import java.util.stream.Collectors;
 
 import lotto.domain.Lotto;
 import lotto.domain.Prize;
 
-public class View {
-
-    private final Scanner scanner;
-
-    public View() {
-        this.scanner = new Scanner(System.in);
-    }
-
-    public int promptExpense() {
-        System.out.println("구입금액을 입력해주세요.");
-        return scanner.nextInt();
-    }
+public class OutputView {
 
     public void printLotto(List<Lotto> lotto) {
         System.out.printf("%d개를 구매했습니다.\n", lotto.size());
@@ -31,27 +17,6 @@ public class View {
         System.out.print("[");
         System.out.print(lotto);
         System.out.println("]");
-    }
-
-    public List<Integer> promptWinningNumbers() {
-        System.out.println("지난 주 당첨 번호를 입력해주세요.");
-        return Arrays.stream(nextNotBlankLine().split(","))
-            .map(String::trim)
-            .map(Integer::parseInt)
-            .collect(Collectors.toList());
-    }
-
-    private String nextNotBlankLine() {
-        String line = scanner.nextLine();
-        while (line.isBlank()) {
-            line = scanner.nextLine();
-        }
-        return line;
-    }
-
-    public int promptBonusNumber() {
-        System.out.println("보너스 볼을 입력해주세요.");
-        return scanner.nextInt();
     }
 
     public void printPrizeHeader() {
@@ -71,4 +36,3 @@ public class View {
         System.out.printf("총 수익률은 %.2f입니다.\n", rate);
     }
 }
-
